Fixed an error in the dotProduct and crossProduct methods. This was apparently not a fatal error, because the program still worked and produced results even before this error was fixed. And fixing the error has not corrected the lack of results on tile 0, huge inflation factor.

+++ BasicTriangle/AbstractPoint.java
@@ -19,8 +19,8 @@ public interface AbstractPoint<T extends AbstractPoint, A extends AbstractAngle>
 
     public boolean colinear(T p);
     public boolean parallel(T p);
-    public float crossProduct(T p);
-    public float dotProduct(T p);
+    public double crossProduct(T p);
+    public double dotProduct(T p);
 
     // it must override equals.
     public boolean equals(Object obj);
